+++ src/store/slices/search/slices/bookmarks/state.ts
@@ -7,3 +7,5 @@ export interface BookmarkItem {
 }
 
 export type BookmarksState = BookmarkItem[];
+
+export const initialState: BookmarksState = [];