+++ src/store/slices/search/slices/entity/state.ts
@@ -29,3 +29,5 @@ export interface InstanceDomainDetailState {
 }
 
 export type EntityState = MetricDetailState | InstanceDetailState | InstanceDomainDetailState | null;
+
+export const initialState: EntityState = null;