+++ src/store/slices/search/slices/history/reducer.ts
@@ -1,9 +1,7 @@
 import { Reducer } from 'redux';
-import { HistoryState } from './state';
+import { HistoryState, initialState } from './state';
 import { HistoryAction } from './actions';
 import { ADD_HISTORY, CLEAR_HISTORY } from './types';
-
-const initialState: HistoryState = [];
 
 const historyReducer: Reducer<HistoryState, HistoryAction> = (state, action) => {
   if (state === undefined) {