+++ src/store/slices/search/slices/history/state.ts
@@ -1,3 +1,5 @@
 import { SearchQuery } from '../../shared/state';
 
 export type HistoryState = SearchQuery[];
+
+export const initialState: HistoryState = [];