+++ src/store/slices/search/slices/view/reducer.ts
@@ -1,9 +1,7 @@
 import { Reducer } from 'redux';
-import { ViewState } from './state';
+import { ViewState, initialState } from './state';
 import { ViewAction } from './actions';
 import { SET_VIEW } from './types';
-
-const initialState: ViewState = ViewState.Index;
 
 const viewReducer: Reducer<ViewState, ViewAction> = (state, action) => {
   if (state === undefined) {