+++ src/components/Aside/Aside.tsx
@@ -6,9 +6,11 @@ import { asideContainer } from './styles';
 import { connect } from 'react-redux';
 import { RootState } from 'store/reducer';
 import { ViewState } from 'store/slices/search/slices/view/state';
+import { EntityType } from 'store/slices/search/shared/state';
 
 const mapStateToProps = (state: RootState) => ({
   view: state.search.view,
+  entity: state.search.entity,
 });
 
 type AsideProps = ReturnType<typeof mapStateToProps>;
@@ -20,19 +22,35 @@ class Aside extends React.Component<AsideProps, {}> {
   }
 
   renderContents() {
-    const { view } = this.props;
+    const { view, entity } = this.props;
     switch (view) {
-      case ViewState.Detail:
-        return (
-          <>
-            <VerticalGroup spacing="lg">
-              <h4>Metrics From Same Namespace</h4>
-              <ReactPlaceholder type="text" rows={6} ready={false}>
-                &nbsp;
-              </ReactPlaceholder>
-            </VerticalGroup>
-          </>
-        );
+      case ViewState.Detail: {
+        if (!entity) {
+          return;
+        }
+        switch (entity.type) {
+          case EntityType.Metric:
+            return (
+              <VerticalGroup spacing="lg">
+                <h4>Metrics From Same Namespace</h4>
+                <ReactPlaceholder type="text" rows={6} ready={false}>
+                  &nbsp;
+                </ReactPlaceholder>
+              </VerticalGroup>
+            );
+          case EntityType.InstanceDomain:
+            return (
+              <VerticalGroup spacing="lg">
+                <h4>Metrics with same Indom</h4>
+                <ReactPlaceholder type="text" rows={6} ready={false}>
+                  &nbsp;
+                </ReactPlaceholder>
+              </VerticalGroup>
+            );
+          default:
+            return;
+        }
+      }
       default:
         return (
           <>