+++ src/pages/Detail/styles.ts
@@ -2,20 +2,6 @@ import { css } from 'emotion';
 
 const detailPageContainer = css`
   grid-area: content;
-  position: relative;
-`;
-
-const detailPageSpinnerContainer = css`
-  position: absolute;
-  top: 0;
-  left: 0;
-  bottom: 0;
-  right: 0;
-  display: flex;
-  width: 100%;
-  height: 100%;
-  justify-content: center;
-  align-items: center;
 `;
 
 const detailPageItem = css`
@@ -92,7 +78,6 @@ const radioBtnGroupContainer = css`
 
 export {
   detailPageContainer,
-  detailPageSpinnerContainer,
   detailPageItem,
   detailPageHeader,
   detailPageTitle,