+++ src/store/store.ts
@@ -1,5 +1,7 @@
 import { createStore, applyMiddleware, compose } from 'redux';
 import thunk from 'redux-thunk';
+import { persistStore } from 'redux-persist';
+
 import rootReducer from './reducer';
 
 const composeEnhancers = (window as any).__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
@@ -8,4 +10,6 @@ const middleware = [thunk];
 
 const store = createStore(rootReducer, {}, composeEnhancers(applyMiddleware(...middleware)));
 
-export { store };
+const persistor = persistStore(store);
+
+export { store, persistor };